requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-RESTX==1.3.0
python-dotenv==1.0.0
google-generativeai==0.3.2
SQLAlchemy==2.0.23